requirements.txt
certifi==2024.2.2
cftime==1.6.3
contourpy==1.2.1
cycler==0.12.1
epop_ephemeris_library @ https://epop.phys.ucalgary.ca/epop_ephemeris_library_1.3.zip#sha256=4970d4e9ce1909c54442169e2984f5524342251056cd7c08eb4132b86211882b
fonttools==4.51.0
h5py==3.11.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
madrigalWeb==3.3
MarkupSafe==2.1.5
matplotlib==3.9.0
mysql-connector-python==8.4.0
netCDF4==1.6.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
spacepy==0.6.0
threadpoolctl==3.5.0
tzdata==2024.1
xgboost==2.0.3
zipp==3.18.1
